splitting body realtions to separate folder
