+++ SPHINXsys/src/for_3D_build/meshes/cell_linked_list.hpp
@@ -6,7 +6,6 @@
 
 #pragma once
 
-#include "body_relation.h"
 #include "base_particles.h"
 #include "base_kernel.h"
 #include "cell_linked_list.h"
