+++ SPHINXsys/src/shared/bodies/base_body.cpp
@@ -7,7 +7,7 @@
 
 #include "sph_system.h"
 #include "base_particles.hpp"
-#include "body_relation.h"
+#include "base_body_relation.h"
 
 namespace SPH
 {
