+++ SPHINXsys/src/shared/body_relations/all_body_relations.h
@@ -1,0 +1,14 @@
+
+#ifndef ALL_BODY_RELATIONS_H
+#define ALL_BODY_RELATIONS_H
+
+/** @file   This is the header file that user code should include to pick up all
+ *          body relation used in SPHinXsys. **/
+
+#pragma once
+
+#include "base_body_relation.h"
+#include "inner_body_relation.h"
+#include "contact_body_relation.h"
+#include "complex_body_relation.h"
+#endif //ALL_BODY_RELATIONS_H
