+++ SPHINXsys/src/shared/particle_dynamics/base_particle_dynamics.h
@@ -34,7 +34,7 @@
 #include "base_data_package.h"
 #include "sph_data_containers.h"
 #include "neighbor_relation.h"
-#include "body_relation.h"
+#include "all_body_relations.h"
 #include "base_body.h"
 
 #include <functional>
