+++ SPHINXsys/src/shared/particle_dynamics/observer_dynamics/observer_dynamics.h
@@ -33,7 +33,7 @@
 #include "all_particle_dynamics.h"
 #include "base_body.h"
 #include "base_particles.h"
-#include "body_relation.h"
+#include "contact_body_relation.h"
 
 namespace SPH
 {
