+++ SPHINXsys/src/shared/particle_dynamics/solid_dynamics/general_solid_dynamics.h
@@ -33,7 +33,7 @@
 #include "all_particle_dynamics.h"
 #include "general_dynamics.h"
 #include "base_kernel.h"
-#include "body_relation.h"
+#include "all_body_relations.h"
 #include "solid_body.h"
 #include "solid_particles.h"
 #include "elastic_solid.h"
